requirements.txt
wandb
hydra-core
pandas
marshmallow_dataclass
transformers
python-dotenv==1.0.0
boto3==1.26.130
torch
gdown
datasets    